extract base functions to utils
